nvlock: support CUDA runtime version 3.1 to 4.2
Support use of nvlock with *unmodified* program using CUDA runtime library.
This has been tested with CUDA driver 4.2 and CUDA runtime 3.1/3.2/4.0/4.2.

Remove program nvlock, as LD_PRELOAD does not work with CUDA 3.x
since the CUDA runtime library uses dlopen() to load the driver.
To use nvlock, prepend its lib/ directory to LD_LIBRARY_PATH.

For CUDA 3.1 and 3.2, monitor cuCtxCreate, cuCtxAttach and cuCtxDetach
to create or destroy context, along with lock or unlock of device.

For CUDA ≥ 4.0, monitor cuCtxSetCurrent, which is used both to set and
unset the current context. CUDA ≥ 4.0 does not use cuCtxCreate, and
instead seems to fills the data structure CUcontext internally.

Test if libnvlock.so is loadable in nvlock program.

To trace execution, configure CMake in RelWithDebInfo or Debug mode.